Minimal render rate support (forceframeupdate)
DOS games and DOSBox tend to only render a frame if something has
changed. As some host enviroments can perform poorly if the app does not
reliably refresh its output regularly (notably Steam overlay) this new
option ensures that DOSBox outputs a minimal amount of frames even if
nothing is changing in the app.

Currently this is only implemented in the OpenGL renderer.

The new config option is under the [render] section: forceframeupdate.
This option can have any of the following values:
* false (default): Disables the feature.
* true: Sets a default minimum frame refresh rate.
* Int: A number representing the desired minimum milliseconds DOSBox
should wait before forcing a refresh. E.g: 50

Note that the rendering system will not force a re-draw of a frame if
the app has rendered a changed frame within the desired minimum refresh.
Thus if the app is outputting 30fps and forceframeupdate is set to 100ms
(10fps) then no duplicate frame will be shown (in theory).

Due to the use of std::stoi for safer conversion of forceframeupdate to
an int the build system has been updated to build with C++11.